move logic to CLI
